feat: setup favorite and list tabs
